+++ toolsrc/include/vcpkg/help.h
@@ -7,13 +7,13 @@
 
 namespace vcpkg::Help
 {
+    extern const CommandStructure COMMAND_STRUCTURE;
+
     void perform_and_exit(const VcpkgCmdArguments& args, const VcpkgPaths& paths);
 
     void help_topic_valid_triplet(const VcpkgPaths& paths);
 
     void print_usage();
-
-    void print_example(const std::string& command_and_arguments);
 
     std::string create_example_string(const std::string& command_and_arguments);
 }